+++ scripts/New-ReleaseBranch.ps1
@@ -2,7 +2,7 @@ Set-StrictMode -Version Latest
 $ErrorActionPreference = "Stop"
 
 $RepoRoot = "T:\VestigantSpotlight"
-$Version = "V1_0_11"
+$Version = "V1_0_13"
 $Branch = "release/$Version"
 Set-Location $RepoRoot
 git checkout main